+++ mycodo/databases/models/output.py
@@ -31,6 +31,7 @@ class Output(CRUDMixin, db.Model):
     # PWM
     pwm_hertz = db.Column(db.Integer, default=None)  # PWM Hertz
     pwm_library = db.Column(db.Text, default=None)  # Library to produce PWM
+    pwm_invert_signal = db.Column(db.Boolean, default=False)  # 90% duty cycle would become 10%
 
     def __repr__(self):
         return "<{cls}(id={s.id})>".format(s=self, cls=self.__class__.__name__)